File upload functionality
File upload functionality added with the help of html5 jquery and
multer. The files are now uploaded, saved to a temp directory so they
can be processed then returned in a response. Once they are returned in
the response they are delted. Will probably need to add more error
checking etc for this process. But so far so good